requirements.txt
accelerate>=0.5.1
albumentations>=1.1.0
clidantic>=0.0.1
jupyter>=1.0.0
pandas>=1.3.0
python-dotenv>=0.19.0
pyyaml>=6.0
rasterio>=1.2.10
seaborn>=0.11.2
tensorboard>=2.7.0
timm>=0.4.12
tqdm>=4.62.0
wheel>=0.37.0

-f https://download.pytorch.org/whl/cu111/torch_stable.html
torch>=1.9.1+cu111
torchvision>=0.10.1+cu111
